+++ xyui/avalonia/gallery/XYUI.Avalonia.Gallery/Views/Density/DensityCoreRulesView.axaml.cs
@@ -1,0 +1,8 @@
+using Avalonia.Controls;
+
+namespace XYUI.Avalonia.Gallery.Views.Density;
+
+public partial class DensityCoreRulesView : UserControl
+{
+    public DensityCoreRulesView() => InitializeComponent();
+}