+++ xyui/avalonia/gallery/XYUI.Avalonia.Gallery/Views/Density/DensityGuardrailsView.axaml.cs
@@ -1,0 +1,8 @@
+using Avalonia.Controls;
+
+namespace XYUI.Avalonia.Gallery.Views.Density;
+
+public partial class DensityGuardrailsView : UserControl
+{
+    public DensityGuardrailsView() => InitializeComponent();
+}